feat: move color functions into utils
